requirements_dev.txt
-e .[testing]
jupyter
ipython
django-extensions
